moved strtobool function so that it can be shared
