+++ sciencebeam_trainer_grobid_tools/utils/string.py
@@ -73,3 +73,20 @@ def parse_dict(expr: str, delimiter: str = '|') -> Dict[str, str]:
         key, value = parse_key_value(fragment)
         d[key] = value
     return d
+
+
+# copied from distutils.util.strtobool
+def strtobool(val: str) -> int:
+    """Convert a string representation of truth to true (1) or false (0).
+
+    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
+    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
+    'val' is anything else.
+    """
+    val = val.lower()
+    if val in ('y', 'yes', 't', 'true', 'on', '1'):
+        return 1
+    elif val in ('n', 'no', 'f', 'false', 'off', '0'):
+        return 0
+    else:
+        raise ValueError("invalid truth value %r" % (val,))
